chore: Secretrefs refactoring, second iteration
