+++ .golangci.yaml
@@ -116,8 +116,6 @@ linters-settings:
         alias: metricsserver
       - pkg: sigs.k8s.io/controller-runtime/pkg/predicate
         alias: ctrlpredicate
-      - pkg: github.com/kyma-project/telemetry-manager/utils/(\w+)
-        alias: utils
       - pkg: github.com/kyma-project/telemetry-manager/internal/reconciler/commonstatus/stubs
         alias: commonStatusStubs
       - pkg: k8s.io/client-go/testing
