+++ .testcoverage.yml
@@ -39,14 +39,14 @@ override:
     path: ^internal/resources/otelcollector$
   - threshold: 78
     path: ^internal/resources/selfmonitor$
+  - threshold: 81
+    path: ^internal/validators/secretref$
   - threshold: 75
     path: ^internal/resourcelock$
   - threshold: 85
     path: ^internal/webhookcert$
   - threshold: 73
     path: ^webhook/logpipeline
-  - threshold: 80
-    path: ^utils$
 
 # Holds regexp rules which will exclude matched files or packages from coverage statistics
 exclude:
