+++ internal/reconciler/metricpipeline/validator.go
@@ -5,7 +5,6 @@ import (
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/validators/tlscert"
-	utils "github.com/kyma-project/telemetry-manager/utils/metricpipeline"
 )
 
 type EndpointValidator interface {
@@ -17,7 +16,7 @@ type TLSCertValidator interface {
 }
 
 type SecretRefValidator interface {
-	Validate(ctx context.Context, refs []telemetryv1alpha1.SecretKeyRef) error
+	ValidateMetricPipeline(ctx context.Context, pipeline *telemetryv1alpha1.MetricPipeline) error
 }
 
 type Validator struct {
@@ -28,7 +27,7 @@ type Validator struct {
 }
 
 func (v *Validator) validate(ctx context.Context, pipeline *telemetryv1alpha1.MetricPipeline) error {
-	if err := v.SecretRefValidator.Validate(ctx, utils.GetSecretRefs(pipeline)); err != nil {
+	if err := v.SecretRefValidator.ValidateMetricPipeline(ctx, pipeline); err != nil {
 		return err
 	}
