+++ internal/reconciler/tracepipeline/validator.go
@@ -5,7 +5,6 @@ import (
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/validators/tlscert"
-	utils "github.com/kyma-project/telemetry-manager/utils/tracepipeline"
 )
 
 type EndpointValidator interface {
@@ -13,7 +12,7 @@ type EndpointValidator interface {
 }
 
 type SecretRefValidator interface {
-	Validate(ctx context.Context, refs []telemetryv1alpha1.SecretKeyRef) error
+	ValidateTracePipeline(ctx context.Context, pipeline *telemetryv1alpha1.TracePipeline) error
 }
 
 type TLSCertValidator interface {
@@ -28,7 +27,7 @@ type Validator struct {
 }
 
 func (v *Validator) validate(ctx context.Context, pipeline *telemetryv1alpha1.TracePipeline) error {
-	if err := v.SecretRefValidator.Validate(ctx, utils.GetSecretRefs(pipeline)); err != nil {
+	if err := v.SecretRefValidator.ValidateTracePipeline(ctx, pipeline); err != nil {
 		return err
 	}
