+++ internal/validators/secretref/secret_ref_test.go
@@ -78,7 +78,7 @@ func TestValidate(t *testing.T) {
 			secretRefValidator := Validator{
 				Client: client,
 			}
-			err := secretRefValidator.Validate(context.TODO(), test.refs)
+			err := secretRefValidator.validate(context.TODO(), test.refs)
 			require.ErrorIs(t, err, test.expectError)
 		})
 	}
@@ -167,6 +167,386 @@ func TestGetValue(t *testing.T) {
 
 			require.Equal(t, test.expectedValue, string(result))
 			require.ErrorIs(t, err, test.expectError)
+		})
+	}
+}
+
+func TestTracePipeline_GetSecretRefs(t *testing.T) {
+	tests := []struct {
+		name         string
+		given        *telemetryv1alpha1.OTLPOutput
+		pipelineName string
+		expected     []telemetryv1alpha1.SecretKeyRef
+	}{
+		{
+			name:         "only endpoint",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Endpoint: telemetryv1alpha1.ValueType{
+					Value: "",
+					ValueFrom: &telemetryv1alpha1.ValueFromSource{
+						SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+							Name: "secret-1",
+							Key:  "endpoint",
+						}},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Key: "endpoint"},
+			},
+		},
+		{
+			name:         "basic auth and header",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Authentication: &telemetryv1alpha1.AuthenticationOptions{
+					Basic: &telemetryv1alpha1.BasicAuthOptions{
+						User: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-1",
+									Namespace: "default",
+									Key:       "user",
+								}},
+						},
+						Password: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-2",
+									Namespace: "default",
+									Key:       "password",
+								}},
+						},
+					},
+				},
+				Headers: []telemetryv1alpha1.Header{
+					{
+						Name: "header-1",
+						ValueType: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-3",
+									Namespace: "default",
+									Key:       "myheader",
+								}},
+						},
+					},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Namespace: "default", Key: "user"},
+				{Name: "secret-2", Namespace: "default", Key: "password"},
+				{Name: "secret-3", Namespace: "default", Key: "myheader"},
+			},
+		},
+		{
+			name:         "basic auth and header (with missing fields)",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Authentication: &telemetryv1alpha1.AuthenticationOptions{
+					Basic: &telemetryv1alpha1.BasicAuthOptions{
+						User: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name: "secret-1",
+									Key:  "user",
+								}},
+						},
+						Password: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Namespace: "default",
+									Key:       "password",
+								}},
+						},
+					},
+				},
+				Headers: []telemetryv1alpha1.Header{
+					{
+						Name: "header-1",
+						ValueType: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-3",
+									Namespace: "default",
+								}},
+						},
+					},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Key: "user"},
+				{Namespace: "default", Key: "password"},
+				{Name: "secret-3", Namespace: "default"},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			sut := telemetryv1alpha1.TracePipeline{ObjectMeta: metav1.ObjectMeta{Name: test.pipelineName}, Spec: telemetryv1alpha1.TracePipelineSpec{Output: telemetryv1alpha1.TracePipelineOutput{OTLP: test.given}}}
+			actual := getSecretRefsTracePipeline(&sut)
+			require.ElementsMatch(t, test.expected, actual)
+		})
+	}
+}
+
+func TestMetricPipeline_GetSecretRefs(t *testing.T) {
+	tests := []struct {
+		name         string
+		given        *telemetryv1alpha1.OTLPOutput
+		pipelineName string
+		expected     []telemetryv1alpha1.SecretKeyRef
+	}{
+		{
+			name:         "only endpoint",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Endpoint: telemetryv1alpha1.ValueType{
+					Value: "",
+					ValueFrom: &telemetryv1alpha1.ValueFromSource{
+						SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+							Name: "secret-1",
+							Key:  "endpoint",
+						}},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Key: "endpoint"},
+			},
+		},
+		{
+			name:         "basic auth and header",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Authentication: &telemetryv1alpha1.AuthenticationOptions{
+					Basic: &telemetryv1alpha1.BasicAuthOptions{
+						User: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-1",
+									Namespace: "default",
+									Key:       "user",
+								}},
+						},
+						Password: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-2",
+									Namespace: "default",
+									Key:       "password",
+								}},
+						},
+					},
+				},
+				Headers: []telemetryv1alpha1.Header{
+					{
+						Name: "header-1",
+						ValueType: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-3",
+									Namespace: "default",
+									Key:       "myheader",
+								}},
+						},
+					},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Namespace: "default", Key: "user"},
+				{Name: "secret-2", Namespace: "default", Key: "password"},
+				{Name: "secret-3", Namespace: "default", Key: "myheader"},
+			},
+		},
+		{
+			name:         "basic auth and header (with missing fields)",
+			pipelineName: "test-pipeline",
+			given: &telemetryv1alpha1.OTLPOutput{
+				Authentication: &telemetryv1alpha1.AuthenticationOptions{
+					Basic: &telemetryv1alpha1.BasicAuthOptions{
+						User: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Namespace: "default",
+									Key:       "user",
+								}},
+						},
+						Password: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name: "secret-2",
+									Key:  "password",
+								}},
+						},
+					},
+				},
+				Headers: []telemetryv1alpha1.Header{
+					{
+						Name: "header-1",
+						ValueType: telemetryv1alpha1.ValueType{
+							Value: "",
+							ValueFrom: &telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+									Name:      "secret-3",
+									Namespace: "default",
+								}},
+						},
+					},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Namespace: "default", Key: "user"},
+				{Name: "secret-2", Key: "password"},
+				{Name: "secret-3", Namespace: "default"},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			sut := telemetryv1alpha1.MetricPipeline{ObjectMeta: metav1.ObjectMeta{Name: test.pipelineName}, Spec: telemetryv1alpha1.MetricPipelineSpec{Output: telemetryv1alpha1.MetricPipelineOutput{OTLP: test.given}}}
+			actual := getSecretRefsMetricPipeline(&sut)
+			require.ElementsMatch(t, test.expected, actual)
+		})
+	}
+}
+
+func TestLogPipeline_GetSecretRefs(t *testing.T) {
+	tests := []struct {
+		name     string
+		given    telemetryv1alpha1.LogPipeline
+		expected []telemetryv1alpha1.SecretKeyRef
+	}{
+		{
+			name: "only variables",
+			given: telemetryv1alpha1.LogPipeline{
+				Spec: telemetryv1alpha1.LogPipelineSpec{
+					Variables: []telemetryv1alpha1.LogPipelineVariableRef{
+						{
+							Name: "password-1",
+							ValueFrom: telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{Name: "secret-1", Key: "password"},
+							},
+						},
+						{
+							Name: "password-2",
+							ValueFrom: telemetryv1alpha1.ValueFromSource{
+								SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{Name: "secret-2", Key: "password"},
+							},
+						},
+					},
+				},
+			},
+
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "secret-1", Key: "password"},
+				{Name: "secret-2", Key: "password"},
+			},
+		},
+		{
+			name: "http output secret refs",
+			given: telemetryv1alpha1.LogPipeline{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "cls",
+				},
+				Spec: telemetryv1alpha1.LogPipelineSpec{
+					Output: telemetryv1alpha1.LogPipelineOutput{
+						HTTP: &telemetryv1alpha1.LogPipelineHTTPOutput{
+							Host: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Name: "creds", Namespace: "default", Key: "host",
+									},
+								},
+							},
+							User: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Name: "creds", Namespace: "default", Key: "user",
+									},
+								},
+							},
+							Password: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Name: "creds", Namespace: "default", Key: "password",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "creds", Namespace: "default", Key: "host"},
+				{Name: "creds", Namespace: "default", Key: "user"},
+				{Name: "creds", Namespace: "default", Key: "password"},
+			},
+		},
+		{
+			name: "http output secret refs (with missing fields)",
+			given: telemetryv1alpha1.LogPipeline{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "cls",
+				},
+				Spec: telemetryv1alpha1.LogPipelineSpec{
+					Output: telemetryv1alpha1.LogPipelineOutput{
+						HTTP: &telemetryv1alpha1.LogPipelineHTTPOutput{
+							Host: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Name: "creds", Namespace: "default",
+									},
+								},
+							},
+							User: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Name: "creds", Key: "user",
+									},
+								},
+							},
+							Password: telemetryv1alpha1.ValueType{
+								ValueFrom: &telemetryv1alpha1.ValueFromSource{
+									SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
+										Namespace: "default", Key: "password",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: []telemetryv1alpha1.SecretKeyRef{
+				{Name: "creds", Namespace: "default"},
+				{Name: "creds", Key: "user"},
+				{Namespace: "default", Key: "password"},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			actual := getSecretRefsLogPipeline(&test.given)
+			require.ElementsMatch(t, test.expected, actual)
 		})
 	}
 }
